POND-19: add description, fix translation, refactor type location
