--- types/cms/CmsImageSlider/NavigationOptions.d.ts
@@ -1,1 +1,0 @@
-export type NavigationOptions = 'none' | 'inside' | 'outside';
