+++ types/cms/SliderNavigationOptions.d.ts
@@ -1,0 +1,1 @@
+export type SliderNavigationOptions = 'none' | 'inside' | 'outside';
